Create REGEX for validate input
