+++ .env.example
@@ -4,4 +4,4 @@ ENV=localhost
 
 # Server host| API HOST
 SERVER_HOST=localhost
-PORT=7001
+PORT=8080
