+++ locales/en.json
@@ -1,10 +1,12 @@
 {
-  "SERVER_ERROR": "Server Error",
-  "VALIDATIONS": {
-    "COMMON": {
-      "authorization": {
-        "required": "Authorization Token is required"
-      }
-    }
-  }
+	"SERVER_ERROR": "Server Error",
+	"VALIDATIONS": {
+		"input": {
+			"required": "Input key is required",
+			"invalid": "Input should contain 0-9 Digits, White Space and Operators (DUP, POP, +, -) only"
+		}
+	},
+	"LOCATION": {
+		"COUNTRY_LIST": "LOCATION.COUNTRY_LIST"
+	}
 }
