+++ package.json
@@ -47,7 +47,6 @@
     "@types/i18n": "^0.8.6",
     "@types/lodash": "^4.14.149",
     "@types/node": "^12.0.8",
-    "@types/uuid": "^3.4.5",
     "@types/validator": "^10.11.1",
     "async": "^3.1.0",
     "body-parser": "^1.19.0",
