+++ src/components/stack-operations/index.ts
@@ -1,1 +1,1 @@
-export { default } from './location.route';
+export { default } from './stackOperations.route';
