--- src/components/stack-operations/models/index.ts
@@ -1,5 +1,0 @@
-import CountryModel from './country.model';
-import StateModel from './state.model';
-import CityModel from './city.model';
-
-export { CountryModel, StateModel, CityModel };
