--- src/components/stack-operations/schemas/index.ts
@@ -1,5 +1,0 @@
-import CountryMaster from './country';
-import StateMaster from './state';
-import CityMaster from './city';
-
-export { CountryMaster, StateMaster, CityMaster };
